backend/ocr/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pytesseract==0.3.8
Pillow==8.3.1
werkzeug==2.0.1
